Fix something

+++ disks/releases/install.sh
@@ -426,7 +426,7 @@ elif [ "${1}" = "patches" ]; then
   echo "BOOTDISK_PHYSDEVPATH=${BOOTDISK_PHYSDEVPATH}"
 
   # NVMe cache handling for models using libsynonvme.so.1
-  rm -f /etc/nvmePorts
+  [ -f /etc/nvmePorts ] && rm -f /etc/nvmePorts
   for P in $(ls -d /sys/block/nvme* 2>/dev/null); do
     if [ -n "${BOOTDISK_PHYSDEVPATH}" -a "${BOOTDISK_PHYSDEVPATH}" = "$(cat ${P}/uevent | grep 'PHYSDEVPATH' | cut -d'=' -f2)" ]; then
       echo "bootloader: ${P}"
@@ -444,7 +444,7 @@ elif [ "${1}" = "patches" ]; then
 
   checkSynoboot
 
-  [ "$(_get_conf_kv supportportmappingv2)" = "yes" ] && dtModel "${2}" || nondtModel "${2}"
+  [ "$(_get_conf_kv supportportmappingv2)" = "yes" ] && dtModel "${2}" || nondtModel "${2}" || true
 
 elif [ "${1}" = "late" ]; then
   echo "Installing addon disks - ${1}"
@@ -503,7 +503,7 @@ elif [ "${1}" = "late" ]; then
   if [ "$(_get_conf_kv supportportmappingv2)" = "yes" ]; then
     echo "Copying /etc.defaults/model.dtb"
     # copy file
-    cp -vf /usr/bin/dtc /tmpRoot/usr/bin/dtc
+    cp -vf dtc /tmpRoot/usr/bin/dtc
     cp -vf /etc/model.dtb /tmpRoot/etc/model.dtb
     cp -vf /etc/model.dtb /tmpRoot/etc.defaults/model.dtb
   else
